requirements.txt
mobase-stubs